Add more abstract component "CardList" to handle responsive Card grids.
