+++ client/src/view/points-sheet/overview/components/TeamCardList.tsx
@@ -1,32 +1,8 @@
 import React from 'react';
-import { Team } from 'shared/dist/model/Team';
-import TeamCard from './TeamCard';
-import { makeStyles, Theme, createStyles } from '@material-ui/core/styles';
 import { Sheet } from 'shared/dist/model/Sheet';
-
-const useStyles = makeStyles((theme: Theme) =>
-  createStyles({
-    root: {
-      display: 'grid',
-      [theme.breakpoints.down('md')]: {
-        gridTemplateColumns: '100%',
-      },
-      [theme.breakpoints.up('md')]: {
-        gridTemplateColumns: 'repeat(auto-fit, calc(100% / 2))',
-      },
-      [theme.breakpoints.up('lg')]: {
-        gridTemplateColumns: 'repeat(auto-fit, calc(100% / 3))',
-      },
-      [theme.breakpoints.up('xl')]: {
-        gridTemplateColumns: 'repeat(auto-fit, calc(100% / 4))',
-      },
-      marginBottom: theme.spacing(2),
-      '& > *': {
-        margin: theme.spacing(1),
-      },
-    },
-  })
-);
+import { Team } from 'shared/dist/model/Team';
+import CardList from '../../../../components/cardlist/CardList';
+import TeamCard from './TeamCard';
 
 interface Props {
   tutorialId: string;
@@ -43,10 +19,8 @@ function TeamCardList({
   onPdfPreviewClicked,
   onGeneratePdfClicked,
 }: Props): JSX.Element {
-  const classes = useStyles();
-
   return (
-    <div className={classes.root}>
+    <CardList>
       {teams.map(team => (
         <TeamCard
           key={team.id}
@@ -57,7 +31,7 @@ function TeamCardList({
           onGeneratePdfClicked={onGeneratePdfClicked}
         />
       ))}
-    </div>
+    </CardList>
   );
 }
